+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 - Modularization supports namespace or locale module
 - Full namespace / locale module support for all file types
+- Support for object and array elements inside json arrays
 
 ### Changed
 - Improve exception handling on IO operations